pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-portfolio_risk_env"
version = "0.1.0"
description = "Portfolio Risk Env environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m portfolio_risk_env.server.app
server = "portfolio_risk_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["portfolio_risk_env", "portfolio_risk_env.server"]
package-dir = { "portfolio_risk_env" = ".", "portfolio_risk_env.server" = "server" }